# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentAskRequest": {
        "properties": {
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "AgentAskRequest",
        "type": "object"
      },
      "AgentFillRequest": {
        "properties": {
          "message": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "AgentFillRequest",
        "type": "object"
      },
      "AgentFillResponse": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentSchema"
          },
          "filled": {
            "description": "Поля, которые помощник заполнил",
            "items": {
              "type": "string"
            },
            "title": "Filled",
            "type": "array"
          },
          "rejected": {
            "description": "Предложенные значения, не прошедшие проверку",
            "items": {
              "$ref": "#/components/schemas/FieldErrorSchema"
            },
            "title": "Rejected",
            "type": "array"
          },
          "reply": {
            "description": "Ответ помощника для показа в чате",
            "title": "Reply",
            "type": "string"
          }
        },
        "required": [
          "reply",
          "filled",
          "rejected",
          "document"
        ],
        "title": "AgentFillResponse",
        "type": "object"
      },
      "AgentTextResponse": {
        "properties": {
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "AgentTextResponse",
        "type": "object"
      },
...
      "DocumentSchema": {
        "properties": {
          "counterparty_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counterparty Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldErrorSchema"
            },
            "title": "Errors",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "missing": {
            "description": "Обязательные поля, которые ещё не заполнены",
            "items": {
              "type": "string"
            },
            "title": "Missing",
            "type": "array"
          },
          "preview": {
            "title": "Preview",
            "type": "string"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "template": {
            "$ref": "#/components/schemas/TemplateSchema"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "unconfirmed": {
            "description": "Распознанное и предложенное агентом без подтверждения",
            "items": {
              "type": "string"
            },
            "title": "Unconfirmed",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "values": {
            "additionalProperties": {
              "$ref": "#/components/schemas/FieldValueSchema"
            },
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "template",
          "counterparty_id",
          "values",
          "errors",
          "missing",
          "unconfirmed",
          "ready",
          "preview",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentSchema",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "description": "Машинный код ошибки, например auth.invalid_token",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "description": "Текст для показа пользователю",
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Подробности ошибки валидации (422)",
            "title": "Errors"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Идентификатор запроса для поиска в логах",
            "title": "Request Id"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FieldErrorSchema": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "key",
          "code",
          "message"
        ],
        "title": "FieldErrorSchema",
        "type": "object"
      },
      "FieldSpecSchema": {
        "properties": {
          "group": {
            "title": "Group",
            "type": "string"
          },
          "hint": {
            "title": "Hint",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "max_length": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Length"
          },
          "required": {
            "title": "Required",
            "type": "boolean"
          },
          "type": {
            "$ref": "#/components/schemas/FieldType"
          }
        },
        "required": [
          "key",
          "label",
          "type",
          "required",
          "group",
          "hint",
          "max_length"
        ],
        "title": "FieldSpecSchema",
        "type": "object"
      },
      "FieldType": {
        "enum": [
          "text",
          "multiline",
          "name",
          "address",
          "email",
          "phone",
          "money",
          "date",
          "integer",
          "inn",
          "kpp",
          "ogrn",
          "bic",
          "account"
        ],
        "title": "FieldType",
        "type": "string"
      },
      "FieldValueSchema": {
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "confirmed": {
            "default": true,
            "title": "Confirmed",
            "type": "boolean"
          },
          "source": {
            "$ref": "#/components/schemas/ValueSource",
            "default": "manual"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value"
        ],
        "title": "FieldValueSchema",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TemplateSchema": {
        "properties": {
          "body_format": {
            "title": "Body Format",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldSpecSchema"
            },
            "title": "Fields",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_builtin": {
            "title": "Is Builtin",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "kind",
          "description",
          "body_format",
          "is_builtin",
          "fields"
        ],
        "title": "TemplateSchema",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValueSource": {
        "description": "Откуда взялось значение. Источник виден в предпросмотре и решает,\nнужно ли подтверждение человека: распознанное и предложенное агентом —\nчерновик, ручной ввод и справочники — нет.",
        "enum": [
          "manual",
          "profile",
          "counterparty",
          "ocr",
          "agent"
        ],
        "title": "ValueSource",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "MAX mini app API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents/{document_id}/agent/ask": {
      "post": {
        "operationId": "agent_ask",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentAskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Ответить на вопрос по документу",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/v1/documents/{document_id}/agent/cover-letter": {
      "post": {
        "operationId": "agent_cover_letter",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Черновик сопроводительного сообщения контрагенту",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/v1/documents/{document_id}/agent/fill": {
      "post": {
        "operationId": "agent_fill",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentFillRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentFillResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Заполнить поля документа из сообщения",
        "tags": [
          "agent"
        ]
      }
    },
...
  }
}